cdc: Add pre-image select generator method
Based on a mutation, creates a pre-image select operation.

Note, this uses raw cql types and obejcts to shortcut parsing etc,
instead of trying to cache by generated query. Hypothesis is that
this is essentially faster.

The routine assumes all rows in a mutation touch same static/regular
columns. If this is not always true it will need additional
calculations.