commitlog: Add optional zero-terminating page to all writes
Fixes #4700

Whe writing a chunk to disk, add a trailing zero-page to make it possible
for commitlog replay to distinguish corrupted recycled segments from those
simply not terminated. We try to write this "atomically" (i.e. in same write
call), but this can of course fail, so "fixes" here is really
"improves/mostly fixes".